Refactor floor plan component to use FloorPlanService for backend interactions and remove unused code
