+++ Frontend/neighbourhub/src/app/services/floor-plan.service.ts
@@ -1,0 +1,85 @@
+import { Injectable } from '@angular/core';
+import { catchError, forkJoin, map, Observable, of, switchMap } from 'rxjs';
+import { UntilDestroy } from '@ngneat/until-destroy';
+import { FloorPlanBackendService } from '../backend/floor-plan-backend.service';
+import { FloorPlan } from '../entities/models/floor-plan.model';
+import { PinPoint } from '../entities/models/pin-point.model';
+
+export type FloorPlanViewModel = FloorPlan & {
+	imageObjectUrl: string | null;
+};
+
+@UntilDestroy()
+@Injectable({
+	providedIn: 'root'
+})
+export class FloorPlanService {
+	private readonly createdObjectUrls: string[] = [];
+
+	constructor(private floorPlanBackendService: FloorPlanBackendService) {}
+
+	public loadFloorPlans(): Observable<FloorPlanViewModel[]> {
+		this.revokeCreatedObjectUrls();
+
+		return this.floorPlanBackendService.getFloorPlans().pipe(
+			map((floorPlans) => [...floorPlans].sort((left, right) => left.floor - right.floor)),
+			switchMap((floorPlans) => {
+				if (floorPlans.length === 0) {
+					return of([] as FloorPlanViewModel[]);
+				}
+
+				return forkJoin(floorPlans.map((floorPlan) => this.loadFloorPlanImage(floorPlan)));
+			})
+		);
+	}
+
+	public uploadFloorPlan(floor: number, image: File): Observable<void> {
+		return this.floorPlanBackendService.uploadFloorPlan(floor, image);
+	}
+
+	public deleteFloorPlan(floorPlanId: string): Observable<void> {
+		return this.floorPlanBackendService.deleteFloorPlan(floorPlanId);
+	}
+
+	public addPinPoint(dto: { latitude: number; longitude: number; title: string; floorPlanId: string }): Observable<void> {
+		return this.floorPlanBackendService.addPinPoint(dto);
+	}
+
+	public deletePinPoint(pinPointId: string): Observable<void> {
+		return this.floorPlanBackendService.deletePinPoint(pinPointId);
+	}
+
+	public dispose(): void {
+		this.revokeCreatedObjectUrls();
+	}
+
+	private loadFloorPlanImage(floorPlan: FloorPlan): Observable<FloorPlanViewModel> {
+		return this.floorPlanBackendService.getFloorPlanImage(floorPlan.id).pipe(
+			map((blob) => ({
+				...floorPlan,
+				imageObjectUrl: this.createObjectUrl(blob)
+			})),
+			catchError((error) => {
+				console.error(`Failed to load floor plan image for floor plan ${floorPlan.id}`, error);
+				return of({
+					...floorPlan,
+					imageObjectUrl: null
+				});
+			})
+		);
+	}
+
+	private createObjectUrl(blob: Blob): string {
+		const objectUrl = window.URL.createObjectURL(blob);
+		this.createdObjectUrls.push(objectUrl);
+		return objectUrl;
+	}
+
+	private revokeCreatedObjectUrls(): void {
+		for (const objectUrl of this.createdObjectUrls) {
+			window.URL.revokeObjectURL(objectUrl);
+		}
+
+		this.createdObjectUrls.length = 0;
+	}
+}
